Break up cabotapp tests into multiple files. Cleanup urls.py imports.
