+++ cabot/cabotapp/tests/test_web.py
@@ -1,0 +1,69 @@
+# -*- coding: utf-8 -*-
+
+from django.core.urlresolvers import reverse
+from django.test.client import Client
+from datetime import timedelta, date
+from cabot.cabotapp.models import Service
+from cabot.cabotapp.views import StatusCheckReportForm
+from .utils import LocalTestCase
+
+
+class TestWebInterface(LocalTestCase):
+
+    def setUp(self):
+        super(TestWebInterface, self).setUp()
+        self.client = Client()
+
+    def test_set_recovery_instructions(self):
+        # Get service page - will get 200 from login page
+        resp = self.client.get(reverse('update-service', kwargs={'pk': self.service.id}), follow=True)
+        self.assertEqual(resp.status_code, 200)
+
+        # Log in
+        self.client.login(username=self.username, password=self.password)
+        resp = self.client.get(reverse('update-service', kwargs={'pk': self.service.id}))
+        self.assertEqual(resp.status_code, 200)
+        self.assertNotIn('username', resp.content)
+
+        snippet_link = 'https://sub.hackpad.com/wiki-7YaNlsC11bB.js'
+        self.assertEqual(self.service.hackpad_id, None)
+        resp = self.client.post(
+            reverse('update-service', kwargs={'pk': self.service.id}),
+            data={
+                'name': self.service.name,
+                'hackpad_id': snippet_link,
+            },
+            follow=True,
+        )
+        self.assertEqual(resp.status_code, 200)
+        reloaded = Service.objects.get(id=self.service.id)
+        self.assertEqual(reloaded.hackpad_id, snippet_link)
+        # Now one on the blacklist
+        blacklist_link = 'https://unapproved_link.domain.com/wiki-7YaNlsC11bB.js'
+        resp = self.client.post(
+            reverse('update-service', kwargs={'pk': self.service.id}),
+            data={
+                'name': self.service.name,
+                'hackpad_id': blacklist_link,
+            },
+            follow=True,
+        )
+        self.assertEqual(resp.status_code, 200)
+        self.assertIn('valid JS snippet link', resp.content)
+        reloaded = Service.objects.get(id=self.service.id)
+        # Still the same
+        self.assertEqual(reloaded.hackpad_id, snippet_link)
+
+    def test_checks_report(self):
+        form = StatusCheckReportForm({
+            'service': self.service.id,
+            'checks': [self.http_check.id],
+            'date_from': date.today() - timedelta(days=1),
+            'date_to': date.today(),
+        })
+        self.assertTrue(form.is_valid())
+        checks = form.get_report()
+        self.assertEqual(len(checks), 1)
+        check = checks[0]
+        self.assertEqual(len(check.problems), 1)
+        self.assertEqual(check.success_rate, 50)
