+++ tox.ini
@@ -2,7 +2,7 @@
 # Runs docker-compose tests. Also creates a development virtualenv.
 # Usage:
 #   tox # runs the test suite
-#   tox -e develop $ create the dev virutalenv
+#   tox -e develop # create the dev virtualenv
 
 [tox]
 envlist = test-mysql, test-psql, flake8
